Remove dstest namespace
The terraform state seems to be a bit messed up
for this namespace.

e.g. RDS instance cloud-platform-27e86c5633608c41
belongs to this namespace, but it's not in the
terraform source code, and doesn't show up in
`terraform state list`, but it's also not showing
up in the orphaned AWS resources list.

I'm deleting the whole namespace to see if that
makes things show up correctly.