Add a DynamoDB table for terraform state locking
This table will be used to lock terraform states for Operations
Engineering projects (initially [this
one](https://github.com/ministryofjustice/github-collaborators).

A later change will add an AWS IAM user with permissions to access both
this dynamodb table, and the S3 bucket (also defined in this folder)
which stores the terraform state files.