Bump prisoner content hub ElasticSearch alarms to 6 minute threshold (#4110)
Following on from https://mojdt.slack.com/archives/C57UPMZLY/p1610615299003400, we increased the available disk space in #4059 and #4060. However, we're still seeing the same alarms. Given the struggle to find a cause, and that they always resolve themselves in exactly 5 minutes, this feels like a good trade-off between getting to the bottom of it and retaining the will to live.