Add formbuilder editor workers service account
The new formbuilder editor in test environment
will be publishing forms into the formbuilder-services-test-* namespaces
and it needs a specific service account to provision new pods applying
k8s files with a kube token.

This service account will be used to provisioning new forms into the
services namespace as we are starting to build the editor
test environment.

So the editor-workers for the test environment can provisioning forms
for draft forms and published forms (formbuilder-services-test-dev and
formbuilder-service-test-production respectively)