Update the namespace deleter script
We now have resources which require the github
terraform provider. Initialising this provider
requires a "token" and "owner" terraform variable.

This change amends the namespace deleter so that
as well as creating a `main.tf`, it also creates a
`variables.tf` and `versions.tf` file from the
templates in the 
`namespace-resources-cli-template/resources`
directory. 

This ensures that all required terraform variables
are declared when the deleter script runs
`terraform apply`.