Enable apply pipeline to skip namespaces
For long-running operations, such as upgrading RDS
instances, we have to pause the apply pipeline for
**every** namespace, so that it doesn't alter the
namespace in which the long-running change is 
happening (which has unpredictable effects).

Repeatedly having to pause/unpause the apply 
pipeline is cumbersome, so this change creates a
mechanism by which individual namespaces can be
skipped.

The `CpEnv::NamespaceDir#apply` method is called
by both the `apply` and `apply-namespace-changes`
pipeline jobs, so the class needs a way to know if
the "skip namespaces" feature should be applied or
not.

This change adds an `enable_skip_namespaces` flag
for that purpose, and sets it to `true` in the 
`apply` script, and `false` in the 
`apply-namespace-changes` script.

With `enable_skip_namespaces` set to true, the
code looks for a file in the top-level namespace
folder called `APPLY_PIPELINE_SKIP_THIS_NAMESPACE`
If this file exists, the namespace is skipped.