Remove minor version from RDS (#4431)
The Terraform module for RDS instances sets the auto minor update to
true as default. This coupled with pinning the minor versions creates a
Terraform error:

```
Error: Error modifying DB Instance cloud-platform-assadasd:
InvalidParameterCombination: Cannot upgrade postgres from 12.5 to 12.3

```

To remedy this, we will have to remove the pinned minor version and give
users the ability to pin in the future. In module 5.13.1 onwards, you
can now set the `auto_minor_version_upgrade` to false and then pin your
version.

https://github.com/ministryofjustice/cloud-platform-terraform-rds-instance/compare/5.13...5.13.1