Create MOJDS accessibility test site (#4439)
* Create MOJDS accessibility test site

The MOJ Design System is going through an accessibility audit. We're creating a demo site for the auditors so that they can test the components in context.

I've put the ownership as OPG and my personal email on it so that we can trace the site back if there are any queries. Since this site isn't for public consumption, I hope that this is acceptable.

* Remove branch protection