Remove http error rate prometheus alert
Note: this is a multi-namespace PR for simplicity, as I'm doing the same thing in 4 namespaces. Let me know if you prefer 4 separate PRs.

Remove the `IncreasedHTTPErrors` prometheus alert as it is becoming increasingly annoying with all the bots scans and probing lately.

We have other, service-side monitoring alerts with Sentry which are more accurate to detect real issues with "real" users (and not bots), so for the time being we are removing the prometheus one.