Set a Ruby version
This sets a recommended version of Ruby to use in this repo with a
`.ruby-version` file which is a common file used by various version
management tools including `rbenv` and `asdf`.

I've chosen 2.6.8 as this matches the version of Ruby which gets
installed as part of the GitHub Actions workflow.