Improve how Ruby version is specified (#5226)
* Set a Ruby version

This sets a recommended version of Ruby to use in this repo with a
`.ruby-version` file which is a common file used by various version
management tools including `rbenv` and `asdf`.

I've chosen 2.6.8 as this matches the version of Ruby which gets
installed as part of the GitHub Actions workflow.

* Upgrade actions/checkout version to 2

This is a newer version of the checkout action which has a few
performance improvements and other features.

* Switch to ruby/setup-ruby action

This replaces `actions/setup-ruby` in the GitHub Actions workflow with
`ruby/setup-ruby` which is the preferred way of setting up Ruby as the
previous action is no longer maintained.

See https://github.com/actions/setup-ruby/pull/99 for more information.

* Enable Bundler cache in GitHub Actions workflow

This takes advantage of a feature of the `ruby/setup-ruby` action which
automatically installs Bundle, the dependencies from a Gemfile and
then caches the installed Gems for the next run. This should improve
performance of the workflow.

* Remove Ruby version from GitHub Actions workflow

Instead, we can get the `ruby/setup-ruby` action to automatically pick
up the version from the `.ruby-version` file and therefore matches the
same version developers will be using locally.

* Ensure `bin-rspec` workflow gets tested

This adds the `bin-rspec` workflow to the list of files which will
trigger it to run, therefore ensuring that any changes to that workflow
will be tested.