Set a maintenance window for hmpps-book-secure-move-frontend-staging
This sets a maintenance window for all the Book a Secure Move databases.
Recently we had an issue where our database was down for a couple of
minutes during its maintenance window which was inadvertently set to a
time when our site is usually highly trafficked.

We've decided to change the maintenance window and set it explicitly to
the early hours of Sunday morning when our site is least used (as moves
aren't booked for Sundays).

Depends on ministryofjustice/cloud-platform-terraform-rds-instance#108
and ministryofjustice/cloud-platform-terraform-elasticache-cluster#25.