Create exported namespace package
After importing this package you can call the function
`namespace.ChangedInPR` which will return all the namespace names
changed in a Pull request.

This is useful for a number of reasons, but more specifically it allows
us to deduplicate the namespace and perform meta actions like ensuring
the PR owner is allowed to make changes to that namespaces.