Add kubeconfig auth to environments repository
Previously this function sat in a package named `authenticate` in the
hoodaw repo. After a discussion with a colleague, it made more sense for
this code to sit here.

It's difficult to test a file grab from S3, so I've just tested the
expected failure if a user doesn't provide the correct creds or inputs.

Test coverage for the whole package after this commit:

```bash
go test -coverprofile=c.out
PASS
coverage: 28.0% of statements
ok
github.com/ministryofjustice/cloud-platform-environments/pkg/authenticate
0.507s
```