Upgrade BaSM UAT database
This environments, although not production, is used frequently for
various reasons and we often see timeouts. It's used by our suppliers to
test our API.

I'm hoping that by upgrading the databases from `t2.small` to
`t3.medium` we'll get a small performance increase that should reduce
the number of issues we see in the environment.