Upgrade BaSM Staging database
This environment, although not production, is used frequently for
various reasons and we often see timeouts. It's used as part of our end
to end tests.

I'm hoping that by upgrading the databases from `t2.small` to
`t3.medium` we'll get a small performance increase that should reduce
the number of issues we see in the environment.