Remove temporary-accommodation dev namespace
This environment was set up on the basis that the new Temporary Accommodation service would have its own separate infrastructure.

We've since agreed, given the similarities, that we're going to try and share the same namespace as the Approved Premises team.

Next steps will include creating a new namespace for both teams "hmpps-community-accommodation-dev" and the eventual retirement of the existing "hmpps-approved-premises-dev" environment.

No apps have been deployed to any of these environments to our knowledge and we have no live users yet. The hope is that we can tear this down as easily as it was set up.