Remove temporary-accommodation prod namespace
This environment was set up on the basis that the new Temporary Accommodation service would have its own separate infrastructure.

We've since agreed, given the similarities, that we're going to try and share the same namespace as the Approved Premises team.

Next steps will include creating a new namespace for both teams "hmpps-community-accommodation-prod".

No apps have been deployed to any of these environments to our knowledge and we have no live users yet.