Create preprod namespace for community accommodation
This is largely a copy and paste from our development environment that's been working well. With some renaming of "dev" to "preprod".

One notable difference is the introduction of resource tagging which is a recent change added to the platform[1]. Tagging our resources should enable read-only console access[2] which may be very useful on preprod for our upcoming pen test.

[1] https://github.com/ministryofjustice/cloud-platform-environments/pull/9113
[2] https://user-guide.cloud-platform.service.justice.gov.uk/documentation/getting-started/accessing-the-cloud-console.html#tagging-all-of-your-namespace-resources