Upgrade `legal-framework-api-staging` db instance class
This temporarily updates the db instance class from `db.t2.small` to
`db.t3.small` ahead of an engine version upgrade.

The current db instance class is deprecated and not supported for PostgreSQL 14.

Before we can upgrade to the recommended instance class (db.t4g.small), we need
to update to a class that is compatible with v11 (i.e t3).

The size of the database has been temporarily upgraded to `large` to speed up
the subsequent database upgrade.