Refactor secrets generation and injection
This is a first attempt to change from a pull, to a push approach for API secrets. As we can have multiple consumers (for now 2, soon 3), using a `map` and `for_each` can DRY the code and make changes more robust.

Note: this should regenerate existing secrets (change from 20 to 32 chars), that is fine and expected.