Refactor Kubernetes secrets in dev
This is because Kubernetes secrets that we updated manually that were
managed by Terraform were being over-written when the pipeline ran
following infrastructure changes. We now store these
Kubernetes secrets separately to prevent this issue.

We have also moved our truststore public key (for mutual TLS) out of
GitHub and into a Kubernetes secret, to remove the dependency on GitHub.