Limit rbac cluster role to view in production namespace
We are assuming this will not break anything, but can't see any other teams doing this.

Is this the correct approach?

We want to limit access to the production namespace so resources can't be easily removed or edited/applied in the CLI with `kubectl`. Still resources are applied via deploy pipeline (service account will have write access).