CAS Pingdom checks are made via the health-kick service
Pingdom is currently unable to detect the CAS services as they aren’t publicly accessible services.

We think there’s two options:

1. Go via the health-kick service which has existing permission to access the health check pages of our service and is used by other services[1]
2. Add all Pingdom IP addresses into each services helm allowlist [2]

On the one hand we are dependant on health-check service being available, on the other we have to maintain a long list of Pingdom IP addresses which could become out of date. Both can generate false positives. On balance we think the health-kick service is worth a go given it’s used by other services already.

[1] https://github.com/ministryofjustice/cloud-platform-environments/blob/6114e5cc0d24d6afc1f2b62c66c496283a9baf3a/namespaces/live.cloud-platform.service.justice.gov.uk/hmpps-manage-adjudications-prod/resources/pingdom.tf#L7
[2] https://github.com/ministryofjustice/hmpps-assessments-api/blob/main/helm_deploy/hmpps-assessments-api/values.yaml#L65