Increase the db storage by 10 percent
This is an aws rds quirk that can be found in their documentation

https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_PIOPS.StorageTypes.html

You can't reduce the amount of storage for a DB instance after storage
as been allocated. When you increase the allocated storage, it must
be by at least 10 percent. If you try to increase the value by less
than 10 percent, you get an error.