Remove repeated service account name
Remove manually-named service account and use the default name.
This service account is used by github, so we don't need to
pay any attention to it.

Keep the custom name in irsa.tf only, as this is the service
account we'll use to talk to AWS from the application running
in the Kubernetes namespace.