Remove resource alerts from datahub test
The CPU alert is triggering all the time but not telling us anything
useful.  We know our non production environments are underspecced
compared to production, and we want to keep it that way to reduce costs.

I think the KubeQuotaExceeded alert is more directly useful as it tells
us whether we are being good citizens of the cloud platform or not.