Increase CCLF UAT RDS read IOPS alert threshold
We are seeing regular alerts because the threshold for the Prometheus
alert for high RDS read IOPS has been set too low (300). A daily batch
job regularly results in ~750 IOPS which leads to false alarms and noisy
alert channels.

This change increases the threshold to 900 which is slightly higher than
what we currently experience but still well below the 3000 IOPS RDS
threshold.