VB-5189 - prisoner retry queue
Our main batch job runs on a per-prison basis. This queue is being
introduced to allow us to capture failed allocation attempts for
individual prisoners, and put them onto a queue for retrying.

without this, we would need to either:
A. retry the whole prison on the main queue.
B. wait until the main batch job re-runs the following day.