Readd string break for consitancy with other deployments
We have had a problem in the past with the length of this key and found
the solution was to add in this escape \". Although nothing is flagging
up I think I'd like to keep it in for consitancy and to make sure it is
covered off and works when in prod.