Add Sentry DSN secret for InfoX Preprod
This commit adds a secret to store the Sentry DSN for InfoX Preprod,
so that it can be moved out of app configuration. Whilst Sentry
DSN values are technically not secret values according to Sentry,
they do allow events to be sent to Sentry without authentication,
so ideally should be kept private.