Create new RDS instance for visit scheduler DB on preprod
After an upgrade to the existing visit scheduler DB, something went wrong
and it is currently stuck in an unusable state. We have to leave the DB
while aws investigate. For now, to unblock the team, we will create a new
rds instance and point the APIs to that, also point the preprod restore job
to it as well, to populate with prod data.